refactor: extract constants & defaults
