+++ lib/module.js
@@ -3,10 +3,7 @@ const glob = require('glob-all')
 
 const logger = require('./logger')
 
-const MODES = {
-  webpack: 'webpack',
-  postcss: 'postcss'
-}
+const { MODES, DEFAULTS } = require('./constants')
 
 function nuxtPurgeCss (moduleOptions) {
   const { srcDir, purgeCSS = {} } = Object.assign({}, this.options)
@@ -15,42 +12,12 @@ function nuxtPurgeCss (moduleOptions) {
 
   logger.start('Loading module')
 
-  const defaults = {
-    mode: MODES.postcss,
-    enabled: this.options.debug ? false : ({ isDev, isClient }) => (!isDev && isClient),
-    paths: [
-      'components/**/*.vue',
-      'layouts/**/*.vue',
-      'pages/**/*.vue',
-      'plugins/**/*.js',
-      'nuxt.config.js'
-    ],
-    styleExtensions: ['.css'],
-    whitelist: ['body', 'html', 'nuxt-progress', '__nuxt', '__layout'],
-    whitelistPatterns: [
-      /-(leave|enter|appear)(|-(to|from|active))$/, // Normal transitions
-      /^nuxt-link(|-exact)-active$/, // Nuxt link classes
-      /^(?!cursor-move).+-move$/, // Move transitions
-      /data-v-.*/ // Keep scoped styles
-    ],
-    whitelistPatternsChildren: [],
-    extractors: [
-      {
-        extractor: (content) => {
-          const contentWithoutStyleBlocks = content.replace(/<style[^]+?<\/style>/gi, '') // Remove inline vue styles
-          return contentWithoutStyleBlocks.match(/[\w-.:/]+(?<!:)/g) || []
-        },
-        extensions: ['html', 'vue', 'js']
-      }
-    ]
-  }
-
   /*
    * Defaults
    */
 
   const mergedConfig = Object.entries(purgeCSS).reduce((options, [key, value]) => {
-    const defaultValue = defaults[key]
+    const defaultValue = DEFAULTS[key]
 
     if (value && typeof value !== 'function' && Array.isArray(defaultValue)) {
       // Merge value with default value if array
@@ -65,7 +32,7 @@ function nuxtPurgeCss (moduleOptions) {
     return options
   }, {})
 
-  const { mode, enabled, ...config } = Object.assign({}, defaults, mergedConfig)
+  const { mode, enabled, ...config } = Object.assign({}, DEFAULTS, mergedConfig)
 
   // transform relative paths
   config.paths = glob.sync(config.paths.map(p => path.join(srcDir, p)))
